requirements.txt
flask
gunicorn
pdf2docx
pdfminer.six
pytesseract
pdf2image
python-docx
Pillow